MWPW-155686 Catalog Milo: Preview Index Follow Up
