+++ .github/workflows/run-preview-index.yaml
@@ -33,7 +33,7 @@ jobs:
       - name: Reindex preview resources
         run: |
           cd .github/preview-index 
-          npm run index
+          npm run full
         env:
           SHAREPOINT_CLIENT_ID: ${{ secrets.SHAREPOINT_CLIENT_ID }}
           SHAREPOINT_TENANT_ID: ${{ secrets.SHAREPOINT_TENANT_ID }}
